+++ packages-exp/auth-exp/test/integration/flows/anonymous.test.ts
@@ -18,6 +18,7 @@
 import { expect, use } from 'chai';
 import * as chaiAsPromised from 'chai-as-promised';
 
+// eslint-disable-next-line import/no-extraneous-dependencies
 import {
   createUserWithEmailAndPassword,
   EmailAuthProvider,
@@ -28,7 +29,6 @@ import {
   updatePassword,
   Auth,
   OperationType
-  // eslint-disable-next-line import/no-extraneous-dependencies
 } from '@firebase/auth-exp';
 import { FirebaseError } from '@firebase/util';