+++ packages/auth/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @firebase/auth
 
+## 0.16.8
+
+### Patch Changes
+
+- [`56a6a9d4a`](https://github.com/firebase/firebase-js-sdk/commit/56a6a9d4af2766154584a0f66d3c4d8024d74ba5) [#5071](https://github.com/firebase/firebase-js-sdk/pull/5071) (fixes [#4932](https://github.com/firebase/firebase-js-sdk/issues/4932)) - Auto initialize `auth-internal` after `auth` has been initialized.
+
 ## 0.16.7
 
 ### Patch Changes