+++ packages/remote-config/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @firebase/remote-config
 
+## 0.1.41
+
+### Patch Changes
+
+- Updated dependencies [[`56a6a9d4a`](https://github.com/firebase/firebase-js-sdk/commit/56a6a9d4af2766154584a0f66d3c4d8024d74ba5)]:
+  - @firebase/component@0.5.4
+  - @firebase/installations@0.4.30
+
 ## 0.1.40
 
 ### Patch Changes